# FullStack/commodity_dashboard/frontend/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Global Commodity Dashboard API",
        "description": "Professional real-time contract-driven API for global commodity tracking.",
        "version": "1.1.0"
    },
    "paths": {
        "/commodities": {
            "get": {
                "summary": "List Commodities",
                "description": "获取全球大宗商品实时报价。前端 SDK 将自动同步此契约。",
                "operationId": "list_commodities_commodities_get",
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "items": {
                                        "$ref": "#/components/schemas/Commodity"
                                    },
                                    "type": "array",
                                    "title": "Response List Commodities Commodities Get"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/commodities/{ticker}": {
            "get": {
                "summary": "Get Commodity",
                "description": "获取单品详情报价。",
                "operationId": "get_commodity_commodities__ticker__get",
                "parameters": [
                    {
                        "name": "ticker",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Ticker"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Commodity"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Commodity": {
                "properties": {
                    "ticker": {
                        "type": "string",
                        "title": "Ticker",
                        "description": "商品代码",
                        "example": "GC=F"
                    },
                    "name": {
                        "type": "string",
                        "title": "Name",
                        "description": "商品全称"
                    },
                    "price": {
                        "type": "number",
                        "exclusiveMinimum": 0.0,
                        "title": "Price",
                        "description": "当前价格"
                    },
                    "unit": {
                        "type": "string",
                        "title": "Unit",
                        "description": "计价单位 (如 oz, bbl)"
                    },
                    "change_pct": {
                        "type": "number",
                        "title": "Change Pct",
                        "description": "当日涨跌幅 (%)"
                    },
                    "category": {
                        "type": "string",
                        "title": "Category",
                        "description": "分类 (如 Precious Metals, Energy)"
                    },
                    "trend": {
                        "type": "string",
                        "title": "Trend",
                        "description": "市场情绪 (Bullish, Bearish, Neutral)",
                        "default": "Neutral"
                    }
                },
                "type": "object",
                "required": [
                    "ticker",
                    "name",
                    "price",
                    "unit",
                    "change_pct",
                    "category"
                ],
                "title": "Commodity"
            },
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
            "ValidationError": {
                "properties": {
                    "loc": {
                        "items": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "integer"
                                }
                            ]
                        },
                        "type": "array",
                        "title": "Location"
                    },
                    "msg": {
                        "type": "string",
                        "title": "Message"
                    },
                    "type": {
                        "type": "string",
                        "title": "Error Type"
                    },
                    "input": {
                        "title": "Input"
                    },
                    "ctx": {
                        "type": "object",
                        "title": "Context"
                    }
                },
                "type": "object",
                "required": [
                    "loc",
                    "msg",
                    "type"
                ],
                "title": "ValidationError"
            }
        }
    }
}